Developer Commands Update
This update moved the developer commands to the correct native of RegisterCommand. Again this will be like the Admin commands update yesterday. Hold off until all the other commands get pushed. If you want to use the old dev resource then delete fxmanifest.lua and rename oldresource.lua to __resource.lua

All the dev commands are working when you have the default fivem chat working. They work along side the admin commands that were pushed yesterday.

With this push came a new event on the inventory script. This is a server side way of adding items to a player. It was needed for the new dev command of giveitem to work.

I have tested this on my dev server and so far have no issues with any of the commands. (Im testing by myself so I can only test so much but was able to test all the commands except for reviveplayer and carryplayer. This also has a new command called getkeys which will give the developer the keys of any vehicle he is currently in.

If you do come across any errors please let me know on here or through the discord. My goal is to push the remaining commands next week (the basic player commands, police commands and ems commands.) once that is done you will no longer need to disable fivem chat.