Fix: fsn_main\money\client
-- Fix event not adding money to players if targetid is nil
-- Updated FSN_dev command 'addmoney' to pass the amount first then target id to match the client side event of adding money to the wallet

This is just a temporary bandaid that should fix events that don't pass the targetid. May put a permanent fix later if I get time. The permanent fix is changing all the client side events to pass the players server id to the event.

The reason this was originally updated was for the new dev command to add money to other players wallets.  Since the command was server sided it already passed the server id of the player so it did not need to GetPlayerServerId again as it would return the wrong value for that command. Haven't actually tested this yet but theoretically this should work with no issues if there are any then let me know.